# openapi.yaml
openapi: 3.1.0
info:
  title: Mutuelle RAG Agent API
  version: 1.0.0
  description: |
    API de l'agent conversationnel RAG pour l'assistance aux adhérents de mutuelle santé.

    **Fonctionnalités principales:**
    - Gestion de sessions de conversation contextuelles
    - Chat avec streaming SSE pour réponses en temps réel
    - Détection d'intention et routage intelligent
    - Récupération RAG (FAQ, garanties, politiques)
    - Escalade vers conseillers humains
    - Ingestion de documents (dev uniquement)

    **Cas d'usage métier:**
    - Informations sur les contrats et garanties
    - Suivi des remboursements
    - Demandes de devis
    - Gestion administrative (attestations, documents)
    - Réclamations et litiges

  contact:
    name: Support API Mutuelle
    email: api-support@mutuelle.fr
  license:
    name: Proprietary
    url: https://mutuelle.fr/legal/api-license
servers:
  - url: https://api.mutuelle.fr/v1
    description: Production
  - url: https://api-staging.mutuelle.fr/v1
    description: Staging
  - url: http://localhost:8080/v1
    description: Development
...
security:
  - BearerAuth: []
paths:
...
  /sessions:
    post:
      tags: [Sessions]
      summary: Créer une nouvelle session de conversation
      description: |
        Initialise une session de conversation pour l'adhérent authentifié.
        Une session regroupe l'historique des échanges et maintient le contexte.
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
            example:
              metadata:
                channel: web
                userAgent: Mozilla/5.0
                referrer: https://mutuelle.fr/espace-adherent
      responses:
        '201':
          description: Session créée avec succès
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
            X-Trace-Id:
              $ref: '#/components/headers/X-Trace-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationSession'
              example:
                sessionId: sess_01HKG7M2V3QN4R5S6T7U8V9W0X
                adherentId: 550e8400-e29b-41d4-a716-446655440001
                status: ACTIVE
                startedAt: '2024-01-25T10:30:00Z'
                lastActivityAt: '2024-01-25T10:30:00Z'
                messageCount: 0
                escalated: false
                metadata:
                  channel: web
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'

    get:
      tags: [Sessions]
      summary: Lister les sessions de l'adhérent
      description: |
        Récupère la liste des sessions de conversation de l'adhérent authentifié.
        Supporte la pagination et le filtrage par statut.
      operationId: listSessions
      parameters:
        - name: status
          in: query
          description: Filtrer par statut de session
          schema:
            $ref: '#/components/schemas/SessionStatus'
        - name: page
          in: query
          description: Numéro de page (0-indexed)
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Taille de la page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          description: Critère de tri
          schema:
            type: string
            enum: [startedAt, lastActivityAt]
            default: lastActivityAt
        - name: direction
          in: query
          description: Direction du tri
          schema:
            type: string
            enum: [ASC, DESC]
            default: DESC
      responses:
        '200':
          description: Liste des sessions
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
            X-Trace-Id:
              $ref: '#/components/headers/X-Trace-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /sessions/{sessionId}:
    get:
      tags: [Sessions]
      summary: Récupérer une session spécifique
      description: Retourne les détails d'une session de conversation
      operationId: getSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Détails de la session
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
            X-Trace-Id:
              $ref: '#/components/headers/X-Trace-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Sessions]
      summary: Fermer une session
      description: |
        Ferme une session de conversation. L'historique est conservé mais
        la session ne peut plus être utilisée pour de nouveaux messages.
      operationId: closeSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '204':
          description: Session fermée avec succès
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
            X-Trace-Id:
              $ref: '#/components/headers/X-Trace-Id'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /sessions/{sessionId}/history:
    get:
      tags: [Sessions]
      summary: Récupérer l'historique de conversation
      description: |
        Retourne tous les messages échangés dans la session.
        Supporte la pagination pour les longues conversations.
      operationId: getSessionHistory
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Historique des messages
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
            X-Trace-Id:
              $ref: '#/components/headers/X-Trace-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageHistoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # CHAT - CONVERSATION AVEC L'AGENT
  # ============================================================================
...
  /ingestion/faq:
    post:
      tags: [Ingestion]
      summary: Injecter des documents FAQ dans le vector store
      description: |
        **Endpoint de développement uniquement** - Ingère des FAQs dans Qdrant pour le RAG.
        Les documents sont chunké, embedé et indexés automatiquement.
      operationId: ingestFaq
      security:
        - BearerAuth: []
      x-dev-only: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FaqIngestionRequest'
            example:
              faqs:
                - id: faq_001
                  question: "Comment modifier mon adresse?"
                  answer: "Pour modifier votre adresse, connectez-vous à votre espace adhérent..."
                  category: ADMINISTRATIF
                  subcategory: "Modification de profil"
                  keywords: ["adresse", "modification", "profil"]
      responses:
        '200':
          description: Ingestion réussie
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
            X-Trace-Id:
              $ref: '#/components/headers/X-Trace-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionResponse'
              example:
                success: true
                message: "Successfully ingested 15 FAQ documents"
                documentsProcessed: 15
                metadata:
                  chunksCreated: 18
                  embeddingsGenerated: 18
                  vectorsIndexed: 18
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /ingestion/guarantees:
    post:
      tags: [Ingestion]
      summary: Injecter des garanties dans le vector store
      description: |
        **Endpoint de développement uniquement** - Ingère des garanties dans Qdrant.
      operationId: ingestGuarantees
      security:
        - BearerAuth: []
      x-dev-only: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuaranteeIngestionRequest'
      responses:
        '200':
          description: Ingestion réussie
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
            X-Trace-Id:
              $ref: '#/components/headers/X-Trace-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /ingestion/policy:
    post:
      tags: [Ingestion]
      summary: Injecter un document de politique dans le vector store
      description: |
        **Endpoint de développement uniquement** - Ingère des documents de politique (CGV, notices, etc.).
        Les documents longs sont automatiquement chunkés.
      operationId: ingestPolicy
      security:
        - BearerAuth: []
      x-dev-only: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyIngestionRequest'
      responses:
        '200':
          description: Ingestion réussie
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
            X-Trace-Id:
              $ref: '#/components/headers/X-Trace-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

# ==============================================================================
# COMPONENTS
# ==============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT contenant les claims:
        - `adherent_id`: UUID de l'adhérent
        - `email`: Email de l'adhérent
        - `roles`: Liste des rôles (USER, ADMIN)
        - `exp`: Date d'expiration
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      description: Identifiant unique de la session de conversation
      schema:
        type: string
        pattern: '^sess_[A-Z0-9]{26}$'
        example: sess_01HKG7M2V3QN4R5S6T7U8V9W0X
  headers:
    X-Request-Id:
      description: Identifiant unique de la requête pour traçabilité
      schema:
        type: string
        format: uuid
        example: req_01HKG8N3W4RP5Q6S7T8U9V0W1X
    X-Trace-Id:
      description: Identifiant de trace distribué (pour corrélation multi-services)
      schema:
        type: string
        example: trace_4a8f2c9d1e3b7f6a
  responses:
    BadRequest:
      description: Requête invalide (validation échouée)
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
        X-Trace-Id:
          $ref: '#/components/headers/X-Trace-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: La requête contient des erreurs de validation
              details:
                field: message
                reason: Le message ne peut pas être vide
            requestId: req_01HKG8N3W4RP5Q6S7T8U9V0W1X
            timestamp: '2024-01-25T10:35:00Z'
    Unauthorized:
      description: Non authentifié (token manquant ou invalide)
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
        X-Trace-Id:
          $ref: '#/components/headers/X-Trace-Id'
        WWW-Authenticate:
          schema:
            type: string
            example: Bearer realm="Mutuelle API"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: AUTHENTICATION_REQUIRED
              message: Token JWT manquant ou invalide
            requestId: req_01HKG8N3W4RP5Q6S7T8U9V0W1X
            timestamp: '2024-01-25T10:35:00Z'
    Forbidden:
      description: Accès interdit (permissions insuffisantes)
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
        X-Trace-Id:
          $ref: '#/components/headers/X-Trace-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INSUFFICIENT_PERMISSIONS
              message: Vous n'avez pas les permissions pour accéder à cette ressource
            requestId: req_01HKG8N3W4RP5Q6S7T8U9V0W1X
            timestamp: '2024-01-25T10:35:00Z'
    NotFound:
      description: Ressource non trouvée
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
        X-Trace-Id:
          $ref: '#/components/headers/X-Trace-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: RESOURCE_NOT_FOUND
              message: La session demandée n'existe pas
              details:
                resourceType: session
                resourceId: sess_01HKG7M2V3QN4R5S6T7U8V9W0X
            requestId: req_01HKG8N3W4RP5Q6S7T8U9V0W1X
            timestamp: '2024-01-25T10:35:00Z'
    RateLimitExceeded:
      description: Limite de taux dépassée
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
        X-Trace-Id:
          $ref: '#/components/headers/X-Trace-Id'
        X-RateLimit-Limit:
          description: Nombre maximum de requêtes autorisées
          schema:
            type: integer
            example: 100
        X-RateLimit-Remaining:
          description: Nombre de requêtes restantes
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          description: Timestamp de réinitialisation du compteur
          schema:
            type: integer
            format: int64
            example: 1706181360
        Retry-After:
          description: Délai avant nouvelle tentative (en secondes)
          schema:
            type: integer
            example: 60
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: RATE_LIMIT_EXCEEDED
              message: Limite de requêtes dépassée
              details:
                limit: 100
                window: 60
                retryAfter: 45
            requestId: req_01HKG8N3W4RP5Q6S7T8U9V0W1X
            timestamp: '2024-01-25T10:35:00Z'
  schemas:
...
    CreateSessionRequest:
      type: object
      properties:
        metadata:
          type: object
          description: Métadonnées contextuelles de la session
          additionalProperties: true
          example:
            channel: web
            userAgent: Mozilla/5.0
            referrer: https://mutuelle.fr/espace-adherent
    ConversationSession:
      type: object
      required:
        - sessionId
        - adherentId
        - status
        - startedAt
        - lastActivityAt
        - messageCount
        - escalated
      properties:
        sessionId:
          type: string
          description: Identifiant unique de la session
          pattern: '^sess_[A-Z0-9]{26}$'
          example: sess_01HKG7M2V3QN4R5S6T7U8V9W0X
        adherentId:
          type: string
          format: uuid
          description: UUID de l'adhérent
        status:
          $ref: '#/components/schemas/SessionStatus'
        startedAt:
          type: string
          format: date-time
          description: Date de création de la session
        lastActivityAt:
          type: string
          format: date-time
          description: Date de dernière activité
        messageCount:
          type: integer
          minimum: 0
          description: Nombre de messages échangés
        currentIntent:
          $ref: '#/components/schemas/Intent'
        escalated:
          type: boolean
          description: Indique si la session a été escaladée
        escalationReason:
          type: string
          nullable: true
          description: Raison de l'escalade
        escalatedAt:
          type: string
          format: date-time
          nullable: true
          description: Date d'escalade
        metadata:
          type: object
          additionalProperties: true
          description: Métadonnées personnalisées
    SessionStatus:
      type: string
      enum:
        - ACTIVE
        - IDLE
        - ESCALATED
        - CLOSED
      description: |
        État de la session:
        - `ACTIVE`: Session active avec échanges récents
        - `IDLE`: Session inactive depuis un certain temps
        - `ESCALATED`: Session transférée à un conseiller humain
        - `CLOSED`: Session fermée définitivement
    SessionListResponse:
      type: object
      required:
        - sessions
        - page
        - totalElements
        - totalPages
      properties:
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/ConversationSession'
        page:
          type: integer
          description: Numéro de page actuelle (0-indexed)
        totalElements:
          type: integer
          description: Nombre total d'éléments
        totalPages:
          type: integer
          description: Nombre total de pages
        size:
          type: integer
          description: Taille de la page

    # ========================================================================
    # CHAT ET MESSAGES
    # ========================================================================
...
    Intent:
      type: string
      enum:
        - CONTRAT
        - REMBOURSEMENT
        - DEVIS
        - ADMINISTRATIF
        - RECLAMATION
        - ESCALATION
        - GENERAL
      description: |
        Intention détectée dans le message utilisateur:
        - `CONTRAT`: Questions sur contrats, garanties, bénéficiaires
        - `REMBOURSEMENT`: Suivi et historique des remboursements
        - `DEVIS`: Demandes de devis et simulations
        - `ADMINISTRATIF`: Modifications admin, documents
        - `RECLAMATION`: Réclamations et litiges
        - `ESCALATION`: Demande de conseiller humain
        - `GENERAL`: Informations générales (FAQ)
...
    MessageHistoryResponse:
      type: object
      required:
        - messages
        - page
        - totalElements
        - totalPages
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatHistoryMessage'
        page:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
...
    FaqIngestionRequest:
      type: object
      required:
        - faqs
      properties:
        faqs:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/FaqItem'
...
    GuaranteeIngestionRequest:
      type: object
      required:
        - guarantees
      properties:
        guarantees:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/GuaranteeItem'
...
    PolicyIngestionRequest:
      type: object
      required:
        - title
        - content
        - docType
        - version
      properties:
        title:
          type: string
          example: "Conditions Générales de Vente 2024"
        content:
          type: string
          minLength: 1
          description: Contenu textuel complet du document
        docType:
          type: string
          example: CGV
          description: Type de document (CGV, NOTICE, etc.)
        version:
          type: string
          example: "2024"
          description: Version du document
        formulaCode:
          type: string
          nullable: true
          example: GOLD
          description: Code formule associée (null = tous)
    IngestionResponse:
      type: object
      required:
        - success
        - message
        - documentsProcessed
      properties:
        success:
          type: boolean
          description: Indique si l'ingestion a réussi
        message:
          type: string
          example: "Successfully ingested 15 FAQ documents"
        documentsProcessed:
          type: integer
          description: Nombre de documents traités
        metadata:
          type: object
          description: Métadonnées d'ingestion
          properties:
            chunksCreated:
              type: integer
              description: Nombre de chunks créés
            embeddingsGenerated:
              type: integer
              description: Nombre d'embeddings générés
            vectorsIndexed:
              type: integer
              description: Nombre de vecteurs indexés

    # ========================================================================
    # ADMIN
    # ========================================================================
...
    ErrorResponse:
      type: object
      required:
        - error
        - requestId
        - timestamp
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
        requestId:
          type: string
          description: ID de la requête pour traçabilité
          example: req_01HKG8N3W4RP5Q6S7T8U9V0W1X
        traceId:
          type: string
          description: ID de trace distribué
          example: trace_4a8f2c9d1e3b7f6a
        timestamp:
          type: string
          format: date-time
          description: Timestamp de l'erreur
...